backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
alembic==1.14.0
pydantic-settings==2.6.1
anthropic==0.40.0
apscheduler==3.10.4
httpx==0.28.0
python-multipart==0.0.12
pytz==2024.2
icalendar==6.1.3
openai==1.57.0
